Clean up the way we retrieve species characters.
We no longer need to separately query for preview characteristics and do
nested queries to get the character name from the preview character.
Instead, as the full character list is being built, we'll just flag
which ones are preview characters and attach the sort order field to
them, so we can extract the right characters to be passed to the
context.

git-svn-id: http://svn.jazkarta.com/newfs-database-api/gobotany/trunk/src@2085 fafe9cc9-101d-4c6e-bffa-87022359e7d1