Try again to improve the Chrome Frame install workflow.
Instead of the overlay, which doesn't seem to be redirecting, try
going directly to the install page. Use two URL parameters: one to
install in single-user mode, which should eliminate the need for
admin privileges and it seems, the need for the user to explicitly
Enable the plug-in; and another to specify a redirect destination,
which does not seem to work reliably but needs further testing.