Try again to improve the flow for installing Chrome Frame on IE.
Make the banner area more noticeable. Make the installer appear 'inline'
and connect it to the banner. Inline mode always prompts to install,
unlike overlay mode which the user can dismiss and it doesn't come
back. Add a note to the banner that asks users to re-start IE after
installing the plug-in.