Fix bug where core search suggestions were no longer appearing.
Ensure the suggestions are stored in lower case in the database.
Add functional tests to alert if these core search suggestions were
to disappear again. Order the suggestions alphabetically in order
to make tests predictable. (Previously the suggestions were left
unordered so as not to favor the top of the alphabet too strongly,
but this actually should not be a problem, and having this new
order is a bit nicer looking and easier to scan as well.)