Fix up the images on the family and genus pages.
Because we never scheduled story work to fully finish these pages
before the first release, we are patching them up for now. The
images shown are the first two species images, because we do not
yet have 'example drawings' marked in the database. The images now
appear in the usual frame and can be enlarged in a lightbox.