Fixes #61: Incorrect back link on pilegroup page.
Since this small template is only ever used in one place, don't bother
with parsing the url or passing in context variables, just get the url
for the top level page.
Also remove the use of "piles_and_urls" and "parent_page" context
variables, which haven't been passed into this template in a while.