Handle missing Family and Genus data.
For now, replace some of the on-the-fly family and genus row creation we
used before, if we didn't import one from our data files that is
referenced by a particular taxon.  This should be removed as soon as we
complete the family and genus data files, as it could have a tendency to
mask any typos in the family, genus, or taxon data.