Level 3 page: 'Change Show Photos Of' menu to native select menu.
The 'Show Photos Of' menu on the Level 3 page was using a widget that did not
behave well on iOS: it did not produce the standard select pane with a list
of values in it, and making a selection required two taps. A native HTML
select menu ought to solve this.