Level 3 page: Change family and genus menus to native select menus.
The family and genus menus were using a combo box widget that brought up a
a keyboard control on mobile devices which, on the iPad, would not dismiss
after a selection was made. With full keyboard entry not truly needed,
simple native select menus ought to work better on mobile devices.