On iOS, make video links open in the device's YouTube app.
Trying to use Shadowbox for videos on iOS is problematic: different behavior
on iOS 4 vs. 5, needing to scroll to the top for the video to play, and
unreliable playback ability (sometimes bumping the scroll helps). So,
use the reliable YouTube app on the device, even though the user will
have to switch back to Safari after watching the video.