Begin restoring unique URLs and the ability to undo via Back button.
These features are no longer working in the new Ember-based code, so they
will need to be repaired. The new ResultsPageState will help set up the
filters and other page elements from a URL hash, and also build the new
URL hash from the state of the filters and page elements.