Move DOM manipulation out into a separate handler.
The FilterController needs to be built early, and causing it to wait on
the DOM to be ready is not necessary and causes other issues with other
pieces of the script which are waiting on the
filter_controller_is_built Deferred.  Move the one-line of DOM
manipulation where we hide the "Loading" message to a separate handler
that waits on the filter controller to be built and ensures the DOM is
ready.