On IE: Allow dismissing the Chrome Frame installation message.
For Issue #102. Some users reported trouble installing the Chrome Frame plug-in.
It was desired to allow the user to view the site without the plug-in anyway,
even though some things might not work. This change adds a link to dismiss the
installation message. The user will be prompted to install again when they
visit the site again in a new browser session.