Capitalize a heading to match other Help-section heading styles.
Second-level page headings throughout the site, especially seen in the Help
section, are normally capitalized in title case. Exceptions on the site
include the top subheading on the Advanced Map page, which is in the form
of a sentence and uses sentence case. Another exception are the plant group
and subgroup headings on the Levels 1 and 2 pages of the simple key, which
are also sentence case because they function more as list items.