Fix some image-scrolling bugs on the species page by updating the HTML.
The new version of SmoothDivScroll we are using asks for less HTML in the initial document,
adding it in dynamically upon activating the control. Our page still had the old HTML,
and when the control added the new HTML, it did so in an odd nested place and all sorts of
havoc ensued with the scrolling, including being able to scroll way past the last image,
icon glitches, and stuck scrolling in some browsers. (#118)