Fix: Glossarizer can now show definitions of upper-case terms.
Previously, terms like FACW (Facultative Wetland) would appear as 'undefined' in
the glossary term popup because the term was being lower-cased before lookup.
Lower-casing is still the right thing to do in order to allow matching of mixed-
case instances of the terms in the pages. This fix just checks the non-lower-case
version as a backup when the definition is not found. (#135)