Fix: Plants with multiple wetland codes no longer show 'not classified.'
Previously, on species pages, plants with multiple wetland indicator codes showed
'not classified.' With this fix, the first of multiple codes is now shown. Any
additional codes are ignored by the importer because wetland indicator categories
represent non-overlapping, mutually exclusive probability ranges, so it does not
appear to be valid to have more than one for each plant. Also, wetland codes and
the text that appears in the user interface have been moved out of the importer
code and into a new CSV file. (#135)