Fix Plant of the Day image cropping on the home page.
Correct the image to be the square large thumbnail, which crops as intended. Center
the image in the left third of the box under the first 'column' above, and add
some vertical padding to give a proportional border and align with heading. Left-align
text to the dotted-line divider above it. Left align the Learn More button to match
other button alignments on the page. (#138)