Remove use of the inline "args" module in our templates.
To avoid having to include the html templates in our build configuration
becaue they define their own inline modules, modify the modules which
require an args object to return a function rather than executing when
parsed.  Then we can simply create a plain old Javascript object in the
html template and pass it into that function as a regular parameter.