For Get More Questions, omit questions that won't currently narrow results.
Switch to a simpler method of returning additional questions, one that takes the
current filtering state into account. Now the best questions are ones that are
likely to be answerable with the current filtering state of the page: questions
with more than one choice selectable, or else length-filter questions. Questions
are still returned in order of ease of observability. (#133)