Home page images: prevent dark border as the images fade.
A longtime minor bug was that a dark border would begin to appear around
some of the home page images because each image was fading in while
keeping the others visible, which caused their semi-transparent edges to
build up. This fix makes the visible image fade out while the next
image fades in. (#150)