Change a style on the species list page and begin using Compass/SASS.
Now the file to edit is sass/styles.scss, from which Compass will generate
the actual CSS used by the site, which is still css/styles.css. Even
though css/styles.css is a generated file, it will be committed to source
control for simplicity in deployment. Reminder comments are at the top
of the styles files, and notes will be added to the README file in the
gobotany-deploy repository. (#142)