Level 3 page: don't support Back button if cookies are disabled.
Cookies are required for our Back button support, but the code wasn't
checking for cookie support, which caused the page to reload at
times it should not have. Now if cookies appear to be turned off,
Back button support is omitted by necessity. In the future we
should enhance the site to notify users where cookies and JS
being turned off will cause things to not work. (#158)