Species list: keep cookies feature disabled after testing it.
The cookies feature restored the last filter state of the page, but it
interferes with our filters specified on the URL: the page can get
into a no-rows state. So now the only support for 'remembering' filters
upon going Back from a species page are with the group and subgroup
filters, which is all this first version aims to handle anyway. If
better memory is desired in the future, we could implement a hash-
based history mechanism where all filter states go on the URL hash. (#142)