Small screens: Fix some problems on Level 3 pages and elsewhere.
Make the intro overlay smaller so the orange button can be seen and pressed.
Omit the plant preview popup for now on small screens because the sizing
of the popup is problematic, so instead go directly to the species page.
Fix some page-width issues I'd accidentally introduced in a prior commit.
(#151)