Slider: no need to support tapping the bar on touch devices.
Remove the code that attempted to move the thumb when tapping on the
bar on touch devices, similar to clicking on the bar on point-and-
click interfaces. Aside from the code not working, evidently this
is not necessary to support on touch devices anyway, as iOS does
not feature it on their native slider control. (#86)