Make the compiled CSS use SASS's compact mode.
Reduce the size of the delivered style sheet and reduce the number
of diff lines in source control upon changes. Restarting 'compass
watch' gets this going. So far in practice during CSS work it
has not been necessary to look at styles.css--using Firebug or
Chrome dev tools and the .scss source seems to be enough. (#174)