Created a working-area dismiss function #166
Instead of everyone who want to dismiss the working area just-in-case
having to check whether the variable is null or not, there is now a
convenience function that dismisses the working area after checking to
see if it is really up.  Also, the working_area variable is now local
to the results.js module scope, instead of being an ugly global stuck
on the App3 object.