Make 'core' Django app self-contained again.
A recent change linked the 'core' Django app to the 'simplekey' app by
a cross-app foreign key reference to the Video model in simplekey. In
order to remove this coupling from core to keep it self-contained, moved
the Video model into core. Rationale is that most of the videos are
about identifying plant groups and are essential content about plants,
so core is an acceptable place to store them. Core also contains other
models not strictly related to plant data (ex. GlossaryTerm, PartnerSite),
so adding Video for now should not be a problem. (#179)