Added little green "Choose this <taxon>" message #190
"Terminal" leads of the dkey, at which you are done with one page and
ready to advance to a new one, looked odd because there was no little
green message above them (like the "See list of 3 species" messages
above leads that go to several taxa).  So this commit adds a little
"Choose this <taxon>" message above the orange button of each terminal
lead so that leads all look visually the same.