Search suggest: return additional suggestions when space allows.
When fewer than the maximum number of suggestions match at the beginning
of the string, look for additional suggestions for the remaining slots
by matching anywhere in the string. For example, if the user starts typing
the word dogwood and gets to 'dog,' there were unfilled slots at the
end of the list that are now filled with suggestions such as
alternate-leaved-dogwood, big-bracted-dogwood, etc. This is similar to
the way Google returns suggestions. (#197)