Search suggest: exclude exact match from suggestions.
For example, if the user has typed the entire word dogwood, exclude the
suggestion dogwood from the suggestions list because, as an exact match,
the user has already typed the entire word and it is of no use to make
it a selection item. This problem always existed but became more
noticeable after adding more generic plant names in #197. (#197)