Importer: improve taxon image filename parsing.
Support underscores in filenames to indicate hyphenated specific epithets. Only
a few images exist with this convention so far, but they are now imported
and using underscores in this manner should be the preferred format from
now on. The rest of the images for plants with hyphenated specific epithets
rely on fallback logic that we've used previously. This logic also, as of
the previous commit, can detect and import var./ssp. names that it couldn't
before. Now the only images for which a taxon is not found during import
are those where the plant name is misspelled in the image filename. (#203)