Start a new app called site in order to begin fresh CSS.
Site can be a global site-wide app representing parts of the Go Botany site
that do not belong to a major feature such as the three keys, PlantShare, etc.
It will have a new base template to be extended by those in the major
features. The aim is to gradually move some things out of simplekey/. (#215)