Species page breadcrumbs, data now obeys ?pile=SLUG
The main /species/ page previously displayed characters for a random
pile if the species were a member of several piles.  Now, you can force
it to display breadcrumbs and characters for a particular pile by giving
it a URL parameter named "pile".  This will let us make the main species
page the only official page for each species.