"species.html" no longer needs last_plant_id_url #217
The variable is provided to the HTML template, but the template never
seems to use it or insert it anywhere.  It might have once been used to
establish the breadcrumbs, but we are now moving on to alternative
means of doing that (currently GET params; perhaps one day a URL hash).