Removed CommonName spamming from search template
Search generation was getting slowed down because we were massively
spamming species pages with their common names, but matching common
names seem to be coming to the tops of search results even without
any spamming.  So we are getting rid of it for now, and if trouble
crops up later then we can explicitly match common names during
sorting instead, like we current do with scientific names.