Replaced friendly-name spamming with boost
There seems to have been an earlier attempt to bring the simple-key
front-page more often into the search results, but at the moment it is
both expensive (it involves spamming the index document) and also does
not work (the page is off the first page of results for the search,
"trees", that is suggested in the comment).  So for the moment I have
simply given the first page more boost.