Search: improve glossary result titles to indicate the page.
In an effort to continue to make our search like other search engines, and
list pages (or anchored content on them) rather than data objects, the
result titles are back to the glossary page titles, but with the term
at the end. Anchors in the URLs are shortened to remove 'term-' for easier
reading. (The names probably will not clash with others in the HTML.)