Improve simple and full key page headings for better orientation.
Make the headings match the link that was clicked to reach the page in
order to reassure and orient the user as they navigate. Do this on
both the main Simple and Full Key pages as well as for the level
below ('Level 2'). This should help the user to better know which
of the two similar-looking keys they are in. (#249)