Simplify switch for showing future navigation links on dev Heroku.
Replace the debug context processor, which involved both the DEBUG setting and
the INTERNAL_IPS setting, with a new variable in our own context processor,
DEV_FEATURES. This will normally be True in local dev, and False in production.
To override for dev Heroku, set an environment variable DEV_FEATURES=True. (#248)