Mock up a way to break up the Full Key remaining non-monocots.
Because the remaining non-monocots in the Full Key amounts to 1,500
plants, that subgroup is too large there to effectively use the
filtering page. This mockup shows the subgroup broken in two based
on leaf arrangement as suggested, which splits the plants about
in half. This suggestion keeps the two parts on the second level
page to avoid another level of click-through for the user. (#254)