Hype the border around dkey selected leads #247
Before, the green border around an already-selected lead was 2px, and
users complained that it was difficult to notice.

Now it is three times thicker (6px), and also we are no longer using
that border color with the pair of leads currently under consideration -
their border just gets the same color as their background. These two
changes together make the border much more noticable. Hopefully.