Show friendly login message when users try accessing certain pages.
When users try to access pages that require login, show a friendly
login message 'Please log in to continue' instead of the red
error message that shows upon a failed login attempt. Also
remove the None response from the test case class, since it
is no longer needed now that redirects are handled. (#235)