Store sighting locations in a more flexible way.
Keep the user's original input string, which can be one of
three ways to enter location, in the database so it will be
available to present again later. But, allow for storing
all the individual pieces of this input as proper
separate fields so it will be easy and efficient to
work with. (#235)