Put leads back where they were for dkey #268
The previous commit moved all child leads beneath *both* parents,
because I got confused during my first read-through of the new, more
detailed dkey design instructions. So this commit puts sub-couplets
right beneath their parent lead.  But as a bonus, so that the project
gets SOMETHING for this pair of commits, my re-visit to the logic here
has resulted in a vastly simpler algorithm that is about half the length
of the original couplet-hierarchy logic.  Hopefully this will, in the
end, help the project going forward.