Begin writing a general-use input-box suggest component.
A component for offering suggestions as the user types in an input
box is needed in several places for PlantShare, including the Sightings
form. Suggestions already are offered for site search. This new
component will be for handling such capabilities site-wide. Most
of its features are working and set up on a test page. (#235)