Focuse the glossarizer solely on text nodes #268
Before, the glossarizer performed a wholesale replacement of all
elements inside of the element that it was directed to glossarize.
Unfortunately this deletion and re-creation of elements ruins any events
that we have already attached to such elements using jQuery.

Now the glossarizer iterates across the nodes inside of the target,
selecting and modifying only the text DOM nodes, so that any elements
present get to have their events and other properties preserved. In
particular, this lets glossarized blocks of text in the dkey keep their
on-click behaviors: figure numbers now correctly pop up their figures,
and the "See list of 20 families in 1b" superscriptions now pop up their
list of taxa.