Fix #271 by correcting "Choose this X" message
Before, "Choose this species" would be displayed above a dkey button
that in fact lead to a tribe, if that tribe in fact contained exactly
one species. Now, we use a more stringent check so that "Choose this X"
messages are only used to decorate dkey buttons that in fact lead to
precisely the taxon that it claims can be chosen.