Improve the keyboard interaction on the suggestions menu.
Previously, scrolling up or down the list would immediately wrap
the selection around to the other end of the list. But, this is
not how HTML5 datalist implementations work--they leave a slot
for the input box, when nothing in the menu is selected, before
wrapping around. So now our box matches that behavior.

Also, the up/down keypresses mistakenly ran even when the menu
was hidden, but now only when it is visible. (#235)