Made a separate list of group names for breadcrumbs #276
Before, the groups as presented in breadcrumb trails had to use the big
long descriptions that are used for them in the pull-down "selectbox" of
groups. Now, they pull their names from the separate "short_group_texts"
dictionary. While the dictionary currently has exactly the same names as
the main dictionary, it can now diverge as we come up with shorter names
for the different groups.