Removed dkey hard-coded Heroku image hostname
The dkey used to always go to one particular Heroku app for its images,
since the main app that Sid was using was not being updated frequently
enough to have the images that were needed. But now each copy of the app
can rely on its own image database - and developers can now start
improving that URL instead of having it always be the same huge slow
one from the older code running on quiet-wind-6510!