Repair bug that paralyzed dkey image loading #279
Our lazy-image-loader only loads images that are currently on-screen,
as determined by the dimensions of the window. The problem is that
images with a "display" CSS value of "none" have coordinates (0,0)
as though they are in the upper-left corner of the screen, and so a
page with hundreds of off-screen but invisible images might trigger
hundreds of image loads if the user is looking at the top of the page.
To try to prevent this we were checking the .style.display of each
image; but this only worked if the image itself was being hidden, but
not if a parent container - like the <figure> that we put around images
in the dkey - was what was getting hidden! So now we are using the more
expensive but accurate ':visible' jQuery filter, and it has fixed things
so that lazy image loading now works correctly on the dkey, instead of
almost freezing the browser as hundreds of invisible images were all
downloaded at the same time.

Also, I accidentally a JS global. Also now fixed.

This is only the first of several steps toward getting #279 all fixed.