Fix #279 with hard limit on number of images
Before, family and genus API calls would return all images for all
species in that family or genus, which was both very slow — in some
cases taking >30 seconds on Heroku and timing out — and which also
tended to overwhelm the browser for a moment once they were actually
delivered. So this commit fixes the situation by applying a hard limit
of 150 images to both API calls. I thought about doing something
fancier, but Elizabeth indicates that the botanists are planning
something bigger (see the comments to #279), so I am going ahead and
fixing this problem an easy and hamfisted way that saves effort going
forward.