Teach the pull-down group menu to do keystrokes #267
Before, clicking on the dkey "Jump to a group" select box would show a
list of groups, but selecting it with the keyboard simply showed the
empty select box content. Now, we intercept keystrokes, not just mouse
clicks, so that the user who is navigating through our form elements
with the keyboard can also access the list of groups.