Fix tab order imposed by recaptcha on the Sign Up form.
Before, the tab order on the Sign Up form was suboptimal, with the
Sign Me Up button taking three more tab presses to get focus after
the user entered the captcha, and it was not even obvious that
this was possible given three extra tab presses needed. Set the
tab order dynamically and the focus to the first field, so users
can sign up more quickly and easily.