Fix #282 by cancelling tooltip if the user clicks
One last quirk remained before #282 was done: if the user moused-over a
figure number and then very quickly clicked before the tooltip-timeout
had even had time to fire, then the click resulted in the Figure pop-up
being shown, THEN the tooltip would appear a moment later over the TOP
of the popup. So the 'click' event on figure numbers needed to manually
cancel any outstanding figure timeout, which this patch achieves by
keeping a reference to the Tooltip for use later by the click event.