Remove block of code that now belongs in "taxa"
This code seems to have been moved to the new "gobotany.taxa"
package, and, as it now stands in the "simplekey/views.py" file,
it lacks necessary imports for most of it even to work. I am therefore
assuming that it was intended to be deleted, and am removing it.