Fix species page breadcrumb paths apropos the Simple and Full keys.
Before, the incorrect key was shown at the base of the breadcrumb path
when using the Full Key but navigating to a species page for a species
belonging to the Simple Key. Now the URLs from the Full Key contain a
request parameter specifying which key the request came from. Also,
separate the notion of the key the user is operating ('key' parameter)
from whether the species is a member of the Simple Key, the latter
now being used to correctly determine whether to show the Full Key
banner on the species page. (#191)