Make Glossarizer into normal JavaScript object #284
The glossarizer used to be an Ember object, even though it has no state
changes that need to be pushed and pulled to the DOM. Now it is a simple
straightforward JS object, and lives in a single file instead of having
its logic split across two files. And, this now allows its "escape()"
function to be exposed where people can use it, which we need for the
work on the dkey.