Fix dkey pages that skip down into a family/genus #284
Before, an "elided" dkey page would load too few images. For example, a
page whose URL names a family, but that shows a key for the genus
because the family only contains one genus, would still load images from
a "family" API URL, and thus only get one image for the whole genus that
the user would then key through. Now, the dkey correctly recognizes if
the HTML has chosen to elide a level of hierarchy, and grabs the correct
set of images.