Species page now survives a missing dkey #238
Before, the species page would die if the species did not appear in the
dkey, or (which is what lead to this bug's discovery) if the dkey had
not yet been imported when a species page was viewed. Now, the species
page will soldier on past a missing dkey entry to at least display.