Fix JS tests for requirejs() 2.1+
Before, the requirejs() function under Node would run its callback
immediately and so we could use it to set variables that refer to the
modules we need. Now, the callback is invoked after a delay to simulate
how browsers work; so we use an escape route that RequireJS provides:
it returns the module as its value, immediately, if it is called with
a single string as its argument.