Moved "test.sh" into "load", removing duplication #264
Before, tox.ini and our little shell script to run our tests had to keep
independent lists of the packages we test, because there was no easy way
from a shell script to grab the list from inside tox.ini. But now that
we can launch the tests command from inside of "load", we can just grab
the list from tox.ini and remove the duplication.