Improve alignments on sightings locator input fields.
Correct the left alignment of the placeholder text. Widen the
boxes a bit, switching to new sprite coordinates in the button
graphics file. Add an option to the suggester control to align
the left edge of the menu to the inside padded edge of the
input box in order to handle menus like these that have wide
rounded sides. (#292)