Make plant name suggestions submit the request upon select.
Before, the code would detect a form with only one non-hidden field and
submit upon select if that were the case, in order to support things like
single search boxes. In this case, there were two non-hidden items due
to the separate input-submit element. So, this is now changed to an
option in the HTML to automatically submit upon selection. (#292)