Remove the sidebar for the Teaching page.
So far the sidebar is used for navigation within a section, as seen in
About and PlantShare, or contextual help information as seen in the
Simple and Full Keys. Because the Teaching page is currently just
one page that does not reside in the About section and does not
relate closely to the keys either, remove the sidebar entirely for
now. (#299)