Refuse to show a species on a dkey genus page #300
Before, we followed the format of Arthur's book by not showing any leads
where there was no longer any choice beneath a genus: when a genus had
only one species. But since species now live off on another page under a
unified "/species/" URL space, we need to generate a fake lead beneath a
single-species genus, that does not actually exist in the book, that
when clicked launches the user over into the "/species/" space.