Add D. Key breadcrumbs to the search engine index.
Add a new test for D. Key breadcrumbs. Remove a test which no longer
applies (species pages no longer necessarily appear as part of the
first page of search results for a family, because the D. Key
breadcrumbs often contain family names and these rank more highly
than the appearance of the family names on the species pages,
likely due to their higher position on the pages. (#195)