Correct and enhance D. Key page titles.
Fix a mistake I had introduced earlier which included the word 'the'
in the main page title. Add ': Dichotomous Key' at the end of all
pages except the main page in order to match our style site-wide,
which is used in our search engine and will be crawled and
displayed by external search engines. (#195)