Abandoned "tox" for an old-fashioned virtualenv
Because "tox" re-installs a package into its virtualenv in order to try
out a new version, the developer cannot simply hit "Save" in their
editor and expect their Django app to magically reload without running
"tox" over again.  So tox turns out not to be suitable for rapid
development.  Plus, because some developers will not have "tox"
installed on their workstation, which meant that we needed virtualenv.py
to sit here in our project anyway, which more or less defeats the point
of having a separate tool to create our virtualenv's anyway.  So.