Make tests insensitive to AWS environment variables
Before, our tests would fail if no AWS variables were set because the
new "plantshare" code creates a real storage class even if it is just
being run as part of our tests.  Now, it is more careful to fall back to
a plain old Storage() class if tests are being run.