Start a separate plantpreview popup for the dkey #298
This plantpreview is an adaptation of the code in the existing
SpeciesSection.js file, but factored out so that it can run
independently of that page.  Currently this new popup can appear when
you click on a dkey species image, but is not yet populated with goodies
like characters and images.  As of yet, the SpeciesSection.js popup is
untouched by this work.