For #319 stop using \u2060 to prevent splitting
Before, we used \u2060 to prevent browsers from splitting a range like
"7-10 mm" right after the en-dash, since that makes one line end with
"7-" and the next start with "10 mm" and puts the initial value in the
range very far from its units. But Windows browsers apparently show a
blank box for this zero-width space character, so we are rescinding it.