Drive species page off of upgraded get_habitats() #219
Before, the Taxon.get_habitats() function did some ugly string parsing
of the old Taxon.habitats field and returned a list of dedicated Habitat
objects for formatting by the search engine; the search template was the
only Taxon code that used .get_habitats().

Now, Taxon.get_habitats() instead returns CharacterValue objects
belonging to the character 'habitat' which, it turns out, is exactly the
same data but sitting in a general-purpose table instead of in the old
dedicated Habitat table.  And now the get_habits() method is used not
only the species index template, but also the species.html template that
runs the /species/ page.