Ensure a UserProfile exists for the user on Avatar upload.
Oops! Before, if we uploaded an avatar but didn't edit any of the
profile fields, the user's profile was never created. Thus, even though
the image is put in the ScreenedImage queue as normal, it was basically
sitting in Limbo - what object would it be attached to when it was
approved? By going ahead and creating a UserProfile the first time the
user uploads an avatar, we avoid this messy condition entirely.