Get 4x increase in glossary blob query time
Requiring the database to order glossary terms, even when we do not
need them ordered, makes the database perform an in-memory quicksort
and seems to quadruple the query time.  This does not affect the speed
of the view as much as I would like, but is a nice first step.