Avoid creating plant divs that never get used
Before, plant <div> elements were created behind the scenes for every
plant in the pile being viewed on the Level 3 page, even though lots
of those species might not belong to the key (usually the Simple Key)
that is currently being viewed.  Now, we are more careful to only
create plant divs for plants that might actually be displayed in
whichever key (simple or full) we are in.