Make the character editor more responsive #226
The previous sequence of JavaScript operations was writing the DOM,
then reading a position or width, then writing the DOM again, then
doing another read, and so forth — requiring the WebKit engine in my
browser to re-render the whole page over and over again, resulting in
a very slow interface that the mouse could easily outrun.  Thanks to
this refactoring, we now do all DOM reads first, then DOM writes.