Rearrange the Sightings page.
Put the Sightings Locator at the top so it will not need
scrolling to see the whole map in an average browser window.
Save space at the top by putting the 'Show recent' label
next to the plant box. Shorten a label. Add anchors so
the two components can be addressed with URL fragment
identifiers #locator and #latest. (#292)