Preserve avatar image aspect ratio.
We still need the avatar to fit inside a fixed-size 84x84 box, but we
don't want to change the image's aspect ratio and get those weird
"stretchy" or "squeezed" effects. Note: this is the "easy" way for now,
without us having to provide a UI that forces users to crop their images into a
square aspect ratio.