Fix sightings marker info window contents.
Allow for an empty list of images. Put the 'more' link on its own
line. Add the plant name to each box to reinforce to the user
what they are looking at when clicking a marker, and to match
the title of the sighting detail page. Italicize the plant
name in a simple way for now. (#292)