Prevent raw TaxonCharacterValue access from admin
Before, an admin could actually visit a raw TCV instance in the admin
interface, which is a bit odd because the TCV table is really more of an
annotated binary relation, not a set of objects that each have a
separate botanic existence.  Also, the page took a minute or so to
display.  And was not useful once it did.  So, it is gone; admins now
create and destroy TCV rows by checking and un-checking the X's in the
new character value editor.