Reorganize Sightings Locator JS modules.
This was necessary for sticking with our established pattern of having
a JS initialization file for each page. Following splitting Sightings
Locator into its own page (no longer on an overall Sightings page with
the list of recent sightings), we already had our resuable UI
component module with that name. Now the resuable code is named with
'Part' and sightings_locator.js is a page-init file. (#292)