Character in admin should show list of values #334
Before, each Character Value was awkwardly accessible as its own kind of
model in the admin interface, leading to several silly consequences; for
example, Django offered a drop-down that let you move a Character Value
around among several different Characters!  Now, there is no way to view
a Character Value as an isolated model; instead, you can only visit a
Character, and are there presented with an inline table of all of the
values defined for that character — a design that maps much more closely
to what the user sees in the user interface, where character values are
displayed in the working area as part of displaying their character.