Change user upload filenames and paths.
Subdivide avatars and sighting photos for ease of organization,
reporting, etc.  Use an md5 hash of the file data to prevent name
collisions - relying on the storage mechanism's auto-increment feature
is inefficient and could cause odd issues with caching, if new files are
given the same name as old, deleted files.  Using a hash should avoid
this except in the case of uploaded the same exact data, in which case
we probably don't care about overwriting the old file.