Correct __name__ attribute on SearchView.
Older versions of haystack required an override of the __name__ instance
method, but this could cause bugs with decorators, or anything that
assumed that __name__ was a string value instead of a method.
See: https://github.com/toastdriven/django-haystack/issues/263

This was discovered while deploying our code to a new Heroku instance.
We had not specified a version number for the newrelic Python plugin,
and it just so happened that the latest version, 1.9.0.21, had added
some code to api.object_wrapper.object_context method that used the
__name__ property, and would cause an exception if the __name__
attribute was instead an instance method.